separate out the changes related to SerializeClosure
