+++ src/hotspot/share/cds/archiveUtils.hpp
@@ -25,7 +25,6 @@
 #ifndef SHARE_CDS_ARCHIVEUTILS_HPP
 #define SHARE_CDS_ARCHIVEUTILS_HPP
 
-#include "cds/serializeClosure.hpp"
 #include "logging/log.hpp"
 #include "memory/iterator.hpp"
 #include "memory/virtualspace.hpp"
