+++ src/hotspot/share/classfile/javaClasses.hpp
@@ -36,7 +36,6 @@
 
 class JvmtiThreadState;
 class RecordComponent;
-class SerializeClosure;
 
 // Interface for manipulating the basic Java classes.
