+++ src/hotspot/share/oops/instanceStackChunkKlass.cpp
@@ -23,7 +23,6 @@
  */
 
 #include "precompiled.hpp"
-#include "cds/serializeClosure.hpp"
 #include "classfile/javaClasses.inline.hpp"
 #include "classfile/vmClasses.hpp"
 #include "compiler/oopMap.inline.hpp"
